fix(console): split catalog class page from instance page + always-visible Open button (G117.2/.4 React-tree port)
In the production React console (products/catalyst/bootstrap/ui), Catalog-tab
and Deployments-tab cards both hard-linked to /app/$id, so the registered
CatalogDetail class page was orphaned and the instance page wrongly rendered the
class instances-list + "+ New instance" button while hiding the Open button. The
correct two-page model lived only in the never-shipped Svelte scaffold
(products/catalyst/console), which hid the bug.

Two distinct pages now:
- CATALOG/class page /catalog/$blueprint (+ provision/chroot variants) →
  Blueprint header + supported topologies + installed-instances list + inline
  "+ New instance" topology-picker dialog. Instance rows link to /app/$id.
- INSTANCE page /app/$id → the one deployed instance, with an always-visible
  "Open" button (silent SSO) when it has an external URL, NO "New instance"
  button. Keeps the current instance's per-cluster fan-out (status.perCluster[]).

Changes:
- AppsPage: AppCard gains isCatalog prop; per-tab link segment (catalog → class
  page, deployments → instance page) with the same chroot/provision scoping.
- AppDetail: drop the class-level InstancesSection + NewInstanceDialog; render a
  small inline per-cluster fan-out (shared PerClusterTable) for THIS instance;
  relabel the Launch button "Launch →" → "Open" (testid btn-launch-app + silent
  SSO unchanged).
- New shared module AppDetail/InstancesSection.tsx exports InstancesSection +
  NewInstanceDialog + PerClusterTable (lifted from AppDetail). Instance rows link
  to /app/$componentId. New instance opens the dialog INLINE (the old
  /catalog/$blueprintName/new link had no route and 404'd) — chose inline dialog
  over adding a /new route: no navigation, reuses the existing component.
- CatalogDetail: delegates the instances list + New-instance to the shared
  InstancesSection.
- router: add /provision/$deploymentId/catalog/$blueprintName (the chroot
  /catalog/$blueprintName already existed).
- Tests: new CatalogDetail.test.tsx + AppsPage per-tab-link + AppDetail
  no-class-content/Open-button tests in the React tree; repaired AppsPage.test
  QueryClientProvider (pre-existing env break) and a stale AppDetail order/title
  assertion. Skipped the dead Svelte-console MOCK blocks in the two g117 e2e
  specs (LIVE-mode walks unaffected).

typecheck + production build green. Full UI suite failures dropped 86 → 73 (all
remaining failures pre-existing on main, unrelated to this change).

Refs #2741 #2743 #2737 #3090
